requirements.txt
asgiref==3.3.4
Django==3.2.4
django-precise-bbcode==1.2.13
djangorestframework==3.12.4
environs==9.3.2
marshmallow==3.12.1
Pillow==8.2.0
psycopg2-binary==2.8.6
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
transliterate==1.10.2
